refactor: extract profile tool handlers from server.ts into src/mcp/tools/
